Split GameMasterButtonPacket into separate packet per subtype
Before this change there were several issues.
First, 0x64 was not showing up in the PacketTypes table for some reason.
Second, GameMasterButtonPacket confusingly referred to the subtype as
both an "action" (in the Action field) and as a "subtype" (in all other fields,
as well as in the PacketTypes table).
Third, the packet format varied by action and so was complex to describe
as a single packet.

Signed-off-by: Dave Thaler <dthaler@microsoft.com>