Lots of improvements found by @NoseyNick
Fixes #99, #100, #101, #103, #104, #106, #110, and #112. Fixes the broken packet table sort and adds the ability to sort by type or name. Changes occurrences of the term "fighter" to "single-seat craft" unless it's in the name of a packet or where the literal term <code>fighter</code> is used in <code>vesselData.xml</code>. More changes to come.